requirements.txt
pefile==1.2.10.post114
python-whois==0.5.2
